# docs/api/openapi.yaml
components:
  schemas:
...
    AccountTypeRequest:
      properties:
        account_type:
          pattern: ^(individual|legal)$
          title: Account Type
          type: string
        actual_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Actual Address
        bank_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Bank Name
        bik:
          anyOf:
          - type: string
          - type: 'null'
          title: Bik
        checking_account:
          anyOf:
          - type: string
          - type: 'null'
          title: Checking Account
        company_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Name
        corr_account:
          anyOf:
          - type: string
          - type: 'null'
          title: Corr Account
        director_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Director Name
        docs_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Docs Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        inn:
          anyOf:
          - type: string
          - type: 'null'
          title: Inn
        kpp:
          anyOf:
          - type: string
          - type: 'null'
          title: Kpp
        legal_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Legal Address
        ogrn:
          anyOf:
          - type: string
          - type: 'null'
          title: Ogrn
      required:
      - account_type
      title: AccountTypeRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    LoginRequest:
      properties:
        email:
          format: email
          title: Email
          type: string
        password:
          title: Password
          type: string
        remember:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Remember
        remember_me:
          default: false
          title: Remember Me
          type: boolean
      required:
      - email
      - password
      title: LoginRequest
      type: object
    LogoutRequest:
      properties:
        refresh_token:
          title: Refresh Token
          type: string
      required:
      - refresh_token
      title: LogoutRequest
      type: object
...
    PasswordChangeBody:
      properties:
        new_password:
          minLength: 8
          title: New Password
          type: string
        old_password:
          minLength: 1
          title: Old Password
          type: string
      required:
      - old_password
      - new_password
      title: PasswordChangeBody
      type: object
    PasswordConfirmRequest:
      properties:
        new_password:
          anyOf:
          - minLength: 8
            type: string
          - type: 'null'
          title: New Password
        password:
          anyOf:
          - minLength: 8
            type: string
          - type: 'null'
          title: Password
        password_confirm:
          anyOf:
          - type: string
          - type: 'null'
          title: Password Confirm
        token:
          title: Token
          type: string
      required:
      - token
      title: PasswordConfirmRequest
      type: object
    PasswordResetRequest:
      properties:
        email:
          format: email
          title: Email
          type: string
      required:
      - email
      title: PasswordResetRequest
      type: object
...
    RefreshRequest:
      properties:
        refresh_token:
          title: Refresh Token
          type: string
      required:
      - refresh_token
      title: RefreshRequest
      type: object
    RegisterRequest:
      properties:
        consents:
          items:
            type: string
          title: Consents
          type: array
        email:
          format: email
          title: Email
          type: string
        password:
          minLength: 8
          title: Password
          type: string
        password_confirm:
          title: Password Confirm
          type: string
      required:
      - email
      - password
      - password_confirm
      title: RegisterRequest
      type: object
    RegisterResponse:
      properties:
        dev_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Dev Code
        email:
          format: email
          title: Email
          type: string
        message:
          title: Message
          type: string
      required:
      - message
      - email
      title: RegisterResponse
      type: object
    RevokeOthersBody:
      properties:
        refresh_token:
          title: Refresh Token
          type: string
      required:
      - refresh_token
      title: RevokeOthersBody
      type: object
...
    TokenResponse:
      properties:
        access_token:
          title: Access Token
          type: string
        refresh_token:
          title: Refresh Token
          type: string
        token_type:
          default: bearer
          title: Token Type
          type: string
      required:
      - access_token
      - refresh_token
      title: TokenResponse
      type: object
...
    TwoFACodeBody:
      properties:
        challenge_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Challenge Token
        code:
          maxLength: 8
          minLength: 6
          title: Code
          type: string
        remember_me:
          default: false
          title: Remember Me
          type: boolean
      required:
      - code
      title: TwoFACodeBody
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    VerifyEmailRequest:
      properties:
        code:
          maxLength: 6
          minLength: 6
          pattern: ^\d{6}$
          title: Code
          type: string
        email:
          format: email
          title: Email
          type: string
      required:
      - email
      - code
      title: VerifyEmailRequest
      type: object
    VerifyEmailResponse:
      properties:
        access_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
        message:
          title: Message
          type: string
        refresh_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
        status:
          title: Status
          type: string
        token_type:
          default: bearer
          title: Token Type
          type: string
      required:
      - message
      - status
      title: VerifyEmailResponse
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: API платформы 3D-моделей для маркетплейсов
  title: KWork Mob API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/auth/2fa/confirm:
    post:
      operationId: owner_2fa_confirm_api_v1_auth_2fa_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFACodeBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Owner 2Fa Confirm
      tags:
      - Аутентификация
  /api/v1/auth/2fa/disable:
    post:
      description: Отключение TOTP (§20.8). Для Owner компании 2FA обязательна — запрещаем.
      operationId: two_fa_disable_api_v1_auth_2fa_disable_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFACodeBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Two Fa Disable
      tags:
      - Аутентификация
  /api/v1/auth/2fa/setup:
    post:
      description: Настройка TOTP для Owner / любого пользователя (§10).
      operationId: owner_2fa_setup_api_v1_auth_2fa_setup_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Owner 2Fa Setup
      tags:
      - Аутентификация
  /api/v1/auth/2fa/status:
    get:
      operationId: owner_2fa_status_api_v1_auth_2fa_status_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Owner 2Fa Status
      tags:
      - Аутентификация
  /api/v1/auth/2fa/verify-login:
    post:
      operationId: verify_login_2fa_api_v1_auth_2fa_verify_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFACodeBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Verify Login 2Fa
      tags:
      - Аутентификация
  /api/v1/auth/account-type:
    post:
      description: 'Выбор типа аккаунта: individual | legal (+ реквизиты юрлица).'
      operationId: account_type_api_v1_auth_account_type_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountTypeRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Account Type
      tags:
      - Аутентификация
  /api/v1/auth/login:
    post:
      description: Вход. Owner с 2FA → challenge; иначе JWT.
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
      tags:
      - Аутентификация
  /api/v1/auth/logout:
    post:
      operationId: logout_api_v1_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Logout
      tags:
      - Аутентификация
  /api/v1/auth/password/change:
    post:
      description: Смена пароля в настройках (§20.8.2).
      operationId: password_change_api_v1_auth_password_change_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Password Change
      tags:
      - Аутентификация
  /api/v1/auth/password/confirm:
    post:
      operationId: password_confirm_api_v1_auth_password_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordConfirmRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Password Confirm
      tags:
      - Аутентификация
  /api/v1/auth/password/forgot:
    post:
      operationId: password_reset_api_v1_auth_password_forgot_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Password Reset
      tags:
      - Аутентификация
  /api/v1/auth/password/reset:
    post:
      operationId: password_reset_api_v1_auth_password_reset_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Password Reset
      tags:
      - Аутентификация
  /api/v1/auth/refresh:
    post:
      operationId: refresh_token_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Refresh Token
      tags:
      - Аутентификация
  /api/v1/auth/register:
    post:
      description: Регистрация по email + пароль + согласия (§2.8).
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register
      tags:
      - Аутентификация
  /api/v1/auth/sessions:
    get:
      description: Активные сессии (refresh-токены) пользователя (§20.8).
      operationId: list_sessions_api_v1_auth_sessions_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: List Sessions
      tags:
      - Аутентификация
  /api/v1/auth/sessions/revoke-others:
    post:
      description: Завершить все сессии, кроме текущей (§20.8).
      operationId: revoke_other_sessions_api_v1_auth_sessions_revoke_others_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeOthersBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Revoke Other Sessions
      tags:
      - Аутентификация
  /api/v1/auth/sessions/{session_id}:
    delete:
      description: Завершить конкретную сессию пользователя (§20.8).
      operationId: revoke_session_api_v1_auth_sessions__session_id__delete
      parameters:
      - in: path
        name: session_id
        required: true
        schema:
          title: Session Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Revoke Session
      tags:
      - Аутентификация
  /api/v1/auth/verify-email:
    post:
      description: Подтверждение email → JWT для выбора типа аккаунта.
      operationId: verify_email_api_v1_auth_verify_email_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyEmailResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Verify Email
      tags:
      - Аутентификация
...